+++ libhfst/src/HfstTokenizer.cc
@@ -92,7 +92,7 @@ const
 
   /* The string begins with a multi character symbol */
   if (multi_char_symbol_end != NULL)
-    { return multi_char_symbol_end - symbol; }
+    { return (int)(multi_char_symbol_end - symbol); }
   if ((128 & *symbol) == 0)
     { return 1; }
   else if ((32 & *symbol) == 0)
